magnolia/requirements.txt
google-genai
python-dotenv
httpx
anthropic
base58
solders
